Cache file / block writing cache fix

+++ crengine/src/hist.cpp
@@ -283,7 +283,7 @@ static void putBookmark( LVStream * stream, CRBookmark * bmk )
     sprintf( percent, "%d.%02d%%", bmk->getPercent()/100, bmk->getPercent()%100 );
     char bmktag[255];
     sprintf(bmktag, "bookmark type=\"%s\" percent=\"%s\" timestamp=\"%d\" shortcut=\"%d\" page=\"%d\"", tname, percent,
-            (int)bmk->getTimestamp(), bmk->getShortcut(), bmk->getBookmarkPage() );
+            (int)bmk->getTimestamp(), (int)bmk->getShortcut(), (int)bmk->getBookmarkPage() );
     putTag(stream, 3, bmktag);
     putTagValue( stream, 4, "start-point", bmk->getStartPos() );
     putTagValue( stream, 4, "end-point", bmk->getEndPos() );
